Added logs to ignore list. Updated testing info in readme.
Changed the DatabaseController to a pointer-to-DatabaseController to avoid unnecessary copying of the DatabaseController struct
Removed unnecessary Print Lines
Changed the JWT expiration time to a global constant
Updated release/debug mode handling in server.go.
Updated naming to prevent ambiguity.
Added testing for the AuthController and environment variables.
Added several mocks for testing.